// let's-build-it-front/src/app/@shared/models/vehicle-model.model.ts
export interface VehicleModelModel {
  id: number;
  externalId: number;
  vehicleProvider?: any;
  sugDegem: string;
  tozeretCd?: any;
  tozeretNm?: any;
  tozeretEretzNm?: any;
  tozar: string;
  degemCd?: any;
  degemNm: string;
  shnatYitzur: number;
  kvuzatAgraCd?: any;
  nefahManoa: number;
  mishkalKolel: number;
  gova?: any;
  hanaaCd?: any;
  hanaaNm: string;
  mazganInd?: any;
  absInd?: any;
  kariotAvirSource?: any;
  misparKariotAvir?: any;
  hegeKoahInd?: any;
  automaticInd?: any;
  halonotHashmalSource?: any;
  misparHalonotHashmal?: any;
  halonBaggInd?: any;
  galgaleySagsogetKalaInd?: any;
  argazInd?: any;
  merkav?: any;
  ramatGimur: string;
  delekCd?: any;
  delekNm?: any;
  misparDlatot: number;
  koahSus: number;
  misparMoshavim?: any;
  bakaratYatzivutInd?: any;
  kosherGriraImBlamim?: any;
  kosherGriraBliBlamim?: any;
  sugTkinaCd?: any;
  sugTkinaNm?: any;
  sugMamirCd?: any;
  sugMamirNm?: any;
  technologiatHanaaCd?: any;
  technologiatHanaaNm?: any;
  kamutCO2?: any;
  kamutNOX?: any;
  kamutPM10?: any;
  kamutHC?: any;
  kamutHCNOX?: any;
  kamutCO?: any;
  kamutCO2City?: any;
  kamutNOXCity?: any;
  kamutPM10City?: any;
  kamutHCCity?: any;
  kamutCOCity?: any;
  kamutCO2Hway?: any;
  kamutNOXHway?: any;
  kamutPM10Hway?: any;
  kamutHCHway?: any;
  kamutCOHway?: any;
  madadYarok?: any;
  kvutzatZihum?: any;
  bakaratStiyaMenativInd?: any;
  bakaratStiyaMenativMakorHatkana?: any;
  niturMerhakMilfanimInd?: any;
  niturMerhakMilfanimMakorHatkana?: any;
  zihuyBeshetahNistarInd?: any;
  bakaratShyutAdaptivitInd?: any;
  zihuyHolcheyRegelInd?: any;
  zihuyHolcheyRegelMakorHatkana?: any;
  maarechetEzerLabalamInd?: any;
  matzlematReverseInd?: any;
  hayshaneyLahatzAvirBatzmigimInd?: any;
  hayshaneyHagorotInd?: any;
  nikudBetihut?: any;
  ramatEivzurBetihuty?: any;
  teuraAutomatitBenesiyaKadimaInd?: any;
  shlitaAutomatitBeorotGvohimInd?: any;
  shlitaAutomatitBeorotGvohimMakorHatkana?: any;
  zihuyMatzavHitkarvutMesukenetInd?: any;
  zihuyTamrureyTnuaInd?: any;
  zihuyTamrureyTnuaMakorHatkana?: any;
  cO2WLTP?: any;
  hcwltp?: any;
  pmwltp?: any;
  noxwltp?: any;
  cowltp?: any;
  cO2WLTPNEDC?: any;
  kinuyMishari: string;
  created: Date;
  updated?: any;
}